+++ SeriesGuide/src/com/battlelancer/seriesguide/ui/ShowsActivity.java
@@ -201,41 +201,6 @@ public class ShowsActivity extends BaseActivity implements AbsListView.OnScrollL
     protected void onStart() {
         super.onStart();
         AnalyticsUtils.getInstance(this).trackPageView("/Shows");
-        onAutoUpdate();
-    }
-
-    /**
-     * Try to launch an delta-update task if certain conditions are met.
-     */
-    private void onAutoUpdate() {
-        // try to run auto-update
-        if (Utils.isNetworkConnected(this)) {
-            final SharedPreferences prefs = PreferenceManager
-                    .getDefaultSharedPreferences(getApplicationContext());
-
-            // check if auto-update is actually enabled
-            final boolean isAutoUpdateEnabled = prefs.getBoolean(
-                    SeriesGuidePreferences.KEY_AUTOUPDATE, true);
-            if (isAutoUpdateEnabled) {
-
-                // check if wifi is required, abort if necessary
-                final boolean isWifiOnly = prefs.getBoolean(
-                        SeriesGuidePreferences.KEY_AUTOUPDATEWLANONLY, true);
-                if (!isWifiOnly || Utils.isWifiConnected(this)) {
-
-                    // only update if at least 15mins have passed since last one
-                    long now = System.currentTimeMillis();
-                    final long previousUpdateTime = prefs.getLong(
-                            SeriesGuidePreferences.KEY_LASTUPDATE, 0);
-                    final boolean isTime = (now - previousUpdateTime) > 15 * DateUtils.MINUTE_IN_MILLIS;
-
-                    if (isTime && !TaskManager.getInstance(this).isUpdateTaskRunning(false)) {
-                        TaskManager.getInstance(this)
-                                .tryUpdateTask(new UpdateTask(false, this), -1);
-                    }
-                }
-            }
-        }
     }
 
     @Override