+++ CHANGELOG.md
@@ -8,16 +8,27 @@ All dates are in the European Central timezone.
 Version 2.7 *(in development)*
 --------------------------------
 
+* FEATURE Quickly check into shows from the main screen.
 * FEATURE Add notification settings for vibrating (thus requiring new VIBRATE permission) and ringtone selection.
-* FEATURE Configuration for widgets: choose between upcoming or recent episodes (only SeriesGuide X), hide watched episodes.
+* FEATURE Configuration for list widgets: choose between upcoming or recent episodes (only SeriesGuide X), hide watched episodes.
+* FEATURE More filter options for activity screen (hide watched, special episodes)
 
 ### Detailed changes:
+
+#### 2.6.14beta (TBA)
+
+* FEATURE Display remaining unseen episodes in overview in portrait mode.
+* TWEAK Auto-update when launching any part of the app.
+* TWEAK Hide episodes with no air date from Recent.
+* FIX Do not display hidden shows on check in screen.
+
 
 #### 2.6.13beta (2012-06-09)
 
 * NOTICE Please re-add your list widgets!
 * FEATURE Quickly check into shows from the main screen.
 * FEATURE More filter options for activity screen (hide watched, special episodes)
+* TWEAK Let most of the AsyncTasks run in parallel to speed up response times.
 * FIX List widget configuration used layout ids which resulted in crashes on new releases.
 * FIX List widgets do not crash the app on an invalid configuration anymore.